chore(app): move simulation source code to module
