+++ app/src/test/java/fr/carbonit/th/simulation/TreasureMapTest.java
@@ -2,10 +2,10 @@ package fr.carbonit.th.simulation;
 
 import fr.carbonit.th.configuration.TreasureHuntConfiguration;
 import fr.carbonit.th.provider.ConfigurationProvider;
-import fr.carbonit.th.simulation.models.Adventurer;
-import fr.carbonit.th.simulation.localisation.Coordinates;
-import fr.carbonit.th.simulation.mappers.TreasureMapMapper;
-import fr.carbonit.th.simulation.models.TreasureMap;
+import fr.carbonit.th.models.Adventurer;
+import fr.carbonit.th.localisation.Coordinates;
+import fr.carbonit.th.mappers.TreasureMapMapper;
+import fr.carbonit.th.models.TreasureMap;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
