+++ simulation/src/main/java/fr/carbonit/th/mappers/AdventurerMapper.java
@@ -1,13 +1,13 @@
-package fr.carbonit.th.simulation.mappers;
+package fr.carbonit.th.mappers;
 
-import fr.carbonit.th.simulation.actions.Action;
-import fr.carbonit.th.simulation.actions.MoveForwardAction;
-import fr.carbonit.th.simulation.actions.TurnLeftAction;
-import fr.carbonit.th.simulation.actions.TurnRightAction;
-import fr.carbonit.th.simulation.models.Adventurer;
-import fr.carbonit.th.simulation.localisation.Coordinates;
-import fr.carbonit.th.simulation.localisation.Orientation;
-import fr.carbonit.th.simulation.localisation.Position;
+import fr.carbonit.th.actions.Action;
+import fr.carbonit.th.actions.MoveForwardAction;
+import fr.carbonit.th.actions.TurnLeftAction;
+import fr.carbonit.th.actions.TurnRightAction;
+import fr.carbonit.th.models.Adventurer;
+import fr.carbonit.th.localisation.Coordinates;
+import fr.carbonit.th.localisation.Orientation;
+import fr.carbonit.th.localisation.Position;
 
 import java.util.LinkedList;
 import java.util.Queue;
